+++ games/feeding-caiso/src/entities/Hazard.js
@@ -1,4 +1,4 @@
-import { GAME_CONFIG } from '/src/utils/Constants.js';
+import { GAME_CONFIG } from '../utils/Constants.js';
 
 export class Hazard {
     constructor(type, x, y, assets) {